+++ commit/__init__.py
@@ -1,3 +1,3 @@
 __all__ = ['core','models','solvers','trk2dictionary']
 from .info import __version__
-from .core import Evaluation
+from .core import Evaluation, setup