+++ commit/core.pyx
@@ -22,21 +22,20 @@ from pkg_resources import get_distribution
 from amico.util import LOG, NOTE, WARNING, ERROR
 
 
-def setup( lmax=12, ndirs=32761 ) :
+def setup( lmax=12, ndirs=None ) :
     """General setup/initialization of the COMMIT framework.
 
     Parameters
     ----------
     lmax : int
         Maximum SH order to use for the rotation phase (default : 12)
-    ndirs : int
-        Number of directions on the half of the sphere representing the possible orientations of the response functions (default : 32761)
+     ndirs : int
+        DEPRECATED. Now, all directions are precomputed.
     """
+    if ndirs is not None:
+        WARNING('"ndirs" parameter is deprecated')
 
-    if not amico.lut.is_valid(ndirs):
-        ERROR( 'Unsupported value for ndirs.\nNote: Supported values for ndirs are [1, 500, 1000, 1500, 2000, 2500, 3000, 3500, 4000, 4500, 5000, 5500, 6000, 6500, 7000, 7500, 8000, 8500, 9000, 9500, 10000, 32761 (default)]' )
-
-    amico.lut.precompute_rotation_matrices( lmax, ndirs )
+    amico.setup( lmax )
 
 
 def load_dictionary_info( filename ):